move package validator to util
