+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject/src/com/google/cloud/tools/eclipse/appengine/newproject/messages.properties
@@ -11,8 +11,6 @@ illegal.service.name=Illegal service name ''{0}'': A service name may contain up
  letters, and hyphens, with no leading or trailing hyphens.
 enter.project.name=Enter a project name.
 project.location.exists: Project location already exists: {0}
-package.ends.with.period: {0} ends with a period.
-package.contains.whitespace: "{0}" contains whitespace.
 app.engine.service=App Engine service:
 add.appengine.standard.facet=Adding App Engine Standard Facet
 app.engine.standard.project.runtimetype=Java version:
