+++ plugins/com.google.cloud.tools.eclipse.util/src/com/google/cloud/tools/eclipse/util/JavaPackageValidator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.eclipse.appengine.newproject;
+package com.google.cloud.tools.eclipse.util;
 
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Status;
