+++ plugins/com.google.cloud.tools.eclipse.util/src/com/google/cloud/tools/eclipse/util/messages.properties
@@ -1,0 +1,2 @@
+package.ends.with.period: {0} ends with a period.
+package.contains.whitespace: "{0}" contains whitespace.
